+++ config/defaults.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/ydb-platform/ydb-go-sdk/v3/balancers"
 	"github.com/ydb-platform/ydb-go-sdk/v3/credentials"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/stack"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xresolver"
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
@@ -84,7 +85,7 @@ func defaultTLSConfig() *tls.Config {
 func defaultConfig() (c *Config) {
 	return &Config{
 		credentials: credentials.NewAnonymousCredentials(
-			credentials.WithSourceInfo("default"),
+			credentials.WithSourceInfo(stack.Record(0)),
 		),
 		balancerConfig: balancers.Default(),
 		tlsConfig:      defaultTLSConfig(),