+++ CHANGELOG.md
@@ -3,9 +3,7 @@
 * Extended metrics (fill database.sql callbacks, recognize TLI error)
 * Refactored config prefix in metrics
 * Removed excess status labels from metrics
-
-## v3.54.3
-* Implement `String` interface for `Driver` struct
+* Implement `fmt.Stringer` interface for `Driver` struct
 
 ## v3.54.2
 * Added context to some internal methods for better tracing