+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Refactored `internal/value/intervalValue.Yql()`
+* The minimum version of Go in `ydb-go-sdk` has been raised to `go1.20`
+
 ## v3.54.3
 * Added per message metadata support for topic api
 * Context for call options now have same lifetime as driver (previous - same lifetime as context for call Open function).